Revert "Класс по управлению базой данных и немного стиля к классу First_circle"
