+++ lib/StartView/StartView.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 import 'CustomPageViewScrollPhisics.dart';
-import '../FirstCircle/First_circle.dart';
+import '../First_circle.dart';
 
 class StartView extends StatefulWidget {
   const StartView({super.key, required this.title});
